requirements.txt
scrapy==2.13.3
scrapy-playwright==0.0.44
playwright==1.54.0
pillow==12.0.0